+++ client/src/index.ejs
@@ -80,9 +80,9 @@
 		</span>
 		<div id="bottomleft-displays">
 			<button id="help-button" tooltip="Help"><img src="<%=require('./img/help.png')%>" tooltip="Help"></button>
-			<div id="dev-chat" class="hide">
+			<!-- <div id="dev-chat" class="hide">
 				<ul id="dev-chat-messages"></ul>
-			</div>
+			</div> -->
 		</div>
 		<div id="bottomright-displays">
 			<div id="chat">
